Refactor multiturn stages into importable functions
